requirements.txt
streamlit
google-cloud-aiplatform
google-api-python-client
Pillow
requests